Rename vips.ts to index.ts and use file: ref for vips
Remove the redundant re-export index.ts and rename vips.ts
in its place. Use file:../vips for the monorepo dependency
instead of a hardcoded prerelease version.
